refactor: exception 위치 및 ErrorCode 수정(#17)
