+++ product/src/main/java/com/sparta/hotdeal/product/infrastructure/repository/ProductRepositoryCustomImpl.java
@@ -3,10 +3,10 @@ package com.sparta.hotdeal.product.infrastructure.repository;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.sparta.hotdeal.product.application.exception.ApplicationException;
+import com.sparta.hotdeal.product.application.exception.ErrorCode;
 import com.sparta.hotdeal.product.domain.entity.product.Product;
 import com.sparta.hotdeal.product.domain.entity.product.QProduct;
-import com.sparta.hotdeal.product.infrastructure.exception.ApplicationException;
-import com.sparta.hotdeal.product.infrastructure.exception.ErrorCode;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import java.util.ArrayList;
